+++ lib/ex_firebase_auth.ex
@@ -14,6 +14,10 @@ defmodule ExFirebaseAuth do
       {ExFirebaseAuth.KeyStore, name: ExFirebaseAuth.KeyStore}
     ]
 
+    if ExFirebaseAuth.Mock.is_enabled?() do
+      ExFirebaseAuth.Mock.generate_and_store_key_pair()
+    end
+
     # See http://elixir-lang.org/docs/stable/elixir/Supervisor.html
     # for other strategies and supported options
     opts = [strategy: :one_for_one, name: ExFirebaseAuth.Supervisor]