+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 
 - Improve performance of fetching public keys by storing them in an ETS table
+- Added `ExFirebaseAuth.Mock` for writing integration tests with ID tokens
 
 ## 0.1.0 Initial Release